third_party/blink/web_tests/fast/html/imports/import-ignore-document-write-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: line 5: Failed to execute 'write' on 'Document': It isn't possible to write into a document from an asynchronously-loaded external script unless it is explicitly opened.
Document.write should be ignored in imports

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

About to call document.write()...
...after calling document.write()
PASS successfullyParsed is true

TEST COMPLETE
